requirements.txt
-e git+git@bitbucket.org:point97/madrona-features.git@HEAD#egg=madrona_features-master
rpc4django
django-tinymce
